fix shmem_file_setup parameter types
Fixing a crash with huge gem mmaps

+++ sys/compat/linuxkpi/common/include/linux/fs.h
@@ -331,7 +331,7 @@ shmem_read_mapping_page(struct address_space *as, int idx)
 	return (shmem_read_mapping_page_gfp(as, idx, 0));
 }
 
-extern struct linux_file *shmem_file_setup(char *name, int size, int flags);
+extern struct linux_file *shmem_file_setup(char *name, loff_t size, unsigned long flags);
 
 static inline void mapping_set_gfp_mask(struct address_space *m, gfp_t mask) {}
 static inline gfp_t mapping_gfp_mask(struct address_space *m)
